use Int

+++ src/snabl/atype.cpp
@@ -3,7 +3,7 @@
 #include "snabl/atype.hpp"
 
 namespace snabl {
-	AType::AType(Lib &lib, Sym id, size_t size):
+	AType::AType(Lib &lib, Sym id, Int size):
 		Def(id), lib(lib), size(size), tag(lib.env.next_type_tag()) { }
 	
 	void AType::call(const Box &val, Pos pos, bool now) const { lib.env.push(val); }
